@@ -1,11 +1,14 @@
-import {Main, H1, H2} from './styled'
-function Logo(){
-  return(
-    <>
-      <Main><H1>Front End</H1>
-      <H2>developer</H2></Main>
-    </>
-  )
+import React from "react";
+import { Main, H1, H2 } from "./styled";
+function Logo() {
+ return (
+  <>
+   <Main>
+    <H1>Front End</H1>
+    <H2>developer</H2>
+   </Main>
+  </>
+ );
 }
 
-export default Logo
+export default Logo;
