@@ -3,14 +3,13 @@ import Navigation from "../../components/nav/navigation";
 import Card from "../../components/card/card";
 import { Align, Alignleft, Alignright, H2 } from "./styled";
 
-export const Projects = () => {
+const Projects = () => {
  return (
   <>
    <Align>
     <Alignleft>
      <Navigation />
     </Alignleft>
-
     <Alignright>
      <H2>Projects</H2>
      <Card />
@@ -19,3 +18,4 @@ export const Projects = () => {
   </>
  );
 };
+export { Projects };
